+++ uni_modules/uni-id/changelog.md
@@ -1,3 +1,7 @@
+## 3.3.8（2021-10-28）
+- 新增 用户账户封禁接口 [详情](https://uniapp.dcloud.net.cn/uniCloud/uni-id?id=ban-account)
+- 新增 用户账户注销接口 [详情](https://uniapp.dcloud.net.cn/uniCloud/uni-id?id=close-account)
+- 修复 未传appid时用户重复注册的Bug
 ## 3.3.7（2021-10-08）
 - 移除部分接口的废弃提示
 ## 3.3.6（2021-09-08）